crm/telegram-bot/requirements.txt
python-telegram-bot[ext]==20.7
aiohttp==3.9.3
python-dotenv==1.0.0
openai==1.12.0
pandas==2.2.1
openpyxl==3.1.2
pdfplumber==0.10.4
Pillow==10.2.0
pytesseract==0.3.10
psycopg2-binary==2.9.9
httpx==0.27.0